Added draggable form and moved action user code code to new files.
