+++ ActionLanguage/ActionEditing/ActionPackEditorForm.cs
@@ -27,7 +27,7 @@ using Conditions;
 
 namespace ActionLanguage
 {
-    public partial class ActionPackEditorForm : Form
+    public partial class ActionPackEditorForm : ExtendedControls.DraggableForm
     {
         ActionFile actionfile;
         List<Tuple<string, string>> events;
@@ -43,7 +43,7 @@ namespace ActionLanguage
         {
             public Panel panel;
             public ExtendedControls.ComboBoxCustom grouptype;
-            public UserControlAB usercontrol;
+            public ActionPackEditBase usercontrol;
         }
 
         List<Group> groups;
@@ -51,13 +51,14 @@ namespace ActionLanguage
         ActionCoreController actioncorecontroller;
         string applicationfolder;
 
-        public ActionPackEditorForm()
+        public ActionPackEditorForm() 
         {
             InitializeComponent();
         }
 
         public delegate List<string> AdditionalNames(string ev);
         public event AdditionalNames onAdditionalNames;             // must set this, provide extra names
+        public Func<string, Condition, ActionPackEditBase> CreateActionPackEdit;    // must set, given a group and condition, what editor do you want?
 
         public void Init(string t, Icon ic, ActionCoreController cp, string appfolder, ActionFile file, List<Tuple<string, string>> ev)
         {
@@ -140,7 +141,7 @@ namespace ActionLanguage
                 g.usercontrol.Dispose();
             }
 
-            g.usercontrol = UserControlAB.Create(g.grouptype.Text);
+            g.usercontrol = CreateActionPackEdit(g.grouptype.Text, c);
             g.usercontrol.Init(c, groupeventlist[g.grouptype.Text], actioncorecontroller, applicationfolder, actionfile, onAdditionalNames , this.Icon);
             ExtendedControls.ThemeableFormsInstance.Instance.ApplyToControls(g.usercontrol, SystemFonts.DefaultFont);
             g.usercontrol.Location = new Point(panelxmargin + 108, 0);
@@ -208,7 +209,7 @@ namespace ActionLanguage
             }
         }
 
-        private void Usercontrol_RemoveEvent(UserControlAB item)
+        private void Usercontrol_RemoveEvent(ActionPackEditBase item)
         {
             foreach (Group g in groups)
             {
@@ -428,96 +429,9 @@ namespace ActionLanguage
 
         #endregion
 
-
-        #region Window Control
-
-        // Mono compatibility
-        private bool _window_dragging = false;
-        private Point _window_dragMousePos = Point.Empty;
-        private Point _window_dragWindowPos = Point.Empty;
-
-        private IntPtr SendMessage(int msg, IntPtr wparam, IntPtr lparam)
+        private void panelVScroll_Resize(object sender, EventArgs e)
         {
-            Message message = Message.Create(this.Handle, msg, wparam, lparam);
-            this.WndProc(ref message);
-            return message.Result;
-        }
-
-        protected override void WndProc(ref Message m)
-        {
-            bool windowsborder = this.FormBorderStyle == FormBorderStyle.Sizable;
-            // Compatibility movement for Mono
-            if (m.Msg == WM.LBUTTONDOWN && (int)m.WParam == 1 && !windowsborder)
-            {
-                int x = unchecked((short)((uint)m.LParam & 0xFFFF));
-                int y = unchecked((short)((uint)m.LParam >> 16));
-                _window_dragMousePos = new Point(x, y);
-                _window_dragWindowPos = this.Location;
-                _window_dragging = true;
-                m.Result = IntPtr.Zero;
-                this.Capture = true;
-            }
-            else if (m.Msg == WM.MOUSEMOVE && (int)m.WParam == 1 && _window_dragging)
-            {
-                int x = unchecked((short)((uint)m.LParam & 0xFFFF));
-                int y = unchecked((short)((uint)m.LParam >> 16));
-                Point delta = new Point(x - _window_dragMousePos.X, y - _window_dragMousePos.Y);
-                _window_dragWindowPos = new Point(_window_dragWindowPos.X + delta.X, _window_dragWindowPos.Y + delta.Y);
-                this.Location = _window_dragWindowPos;
-                this.Update();
-                m.Result = IntPtr.Zero;
-            }
-            else if (m.Msg == WM.LBUTTONUP)
-            {
-                _window_dragging = false;
-                _window_dragMousePos = Point.Empty;
-                _window_dragWindowPos = Point.Empty;
-                m.Result = IntPtr.Zero;
-                this.Capture = false;
-            }
-            // Windows honours NCHITTEST; Mono does not
-            else if (m.Msg == WM.NCHITTEST)
-            {
-                base.WndProc(ref m);
-
-                if ((int)m.Result == HT.CLIENT)
-                {
-                    int x = unchecked((short)((uint)m.LParam & 0xFFFF));
-                    int y = unchecked((short)((uint)m.LParam >> 16));
-                    Point p = PointToClient(new Point(x, y));
-
-                    if (p.X > this.ClientSize.Width - statusStripCustom.Height && p.Y > this.ClientSize.Height - statusStripCustom.Height)
-                    {
-                        m.Result = (IntPtr)HT.BOTTOMRIGHT;
-                    }
-                    else if (p.Y > this.ClientSize.Height - statusStripCustom.Height)
-                    {
-                        m.Result = (IntPtr)HT.BOTTOM;
-                    }
-                    else if (p.X > this.ClientSize.Width - 5)       // 5 is generous.. really only a few pixels gets thru before the subwindows grabs them
-                    {
-                        m.Result = (IntPtr)HT.RIGHT;
-                    }
-                    else if (p.X < 5)
-                    {
-                        m.Result = (IntPtr)HT.LEFT;
-                    }
-                    else if (!windowsborder)
-                    {
-                        m.Result = (IntPtr)HT.CAPTION;
-                    }
-                }
-            }
-            else
-            {
-                base.WndProc(ref m);
-            }
-        }
-
-        private void label_index_MouseDown(object sender, MouseEventArgs e)
-        {
-            ((Control)sender).Capture = false;
-            SendMessage(WM.NCLBUTTONDOWN, (IntPtr)HT.CAPTION, IntPtr.Zero);
+            PositionGroups(false);
         }
 
         private void panel_minimize_Click(object sender, EventArgs e)
@@ -525,362 +439,15 @@ namespace ActionLanguage
             this.WindowState = FormWindowState.Minimized;
         }
 
-        private void panel_close_Click(object sender, EventArgs e)
+        public void panel_close_Click(object sender, EventArgs e)
         {
             Close();
         }
 
-        #endregion
-
-        private void panelVScroll_Resize(object sender, EventArgs e)
+        private void label_index_MouseDown(object sender, MouseEventArgs e)
         {
-            PositionGroups(false);
+            DragMoveMode((Control)sender);
         }
 
-    }
-
-    /// <summary>
-    /// /////////////////////////////////////////////////////////////////////////////////////
-    /// </summary>
-
-    public class UserControlAB : UserControl
-    {
-        public Condition cd;
-        
-        public ActionFile actionfile;
-
-        public ActionCoreController actioncorecontroller;
-        public string applicationfolder;
-        public System.Drawing.Icon Icon;
-
-        public System.Action<UserControlAB> RemoveItem;
-        public System.Action RefreshEvent;
-
-        public const int panelxmargin = 3;
-        public const int panelymargin = 1;
-
-        static public UserControlAB Create(string groupname)
-        {
-            UserControlAB ucab;
-            ucab = new UserControlProgram();
-            return ucab;
-        }
-
-        protected ActionPackEditorForm.AdditionalNames onAdditionalNames;
-
-        public void Init(Condition cond, List<string> events, ActionCoreController cp, string appfolder, ActionFile file , ActionPackEditorForm.AdditionalNames func , Icon ic)
-        {
-            Icon = ic;
-            onAdditionalNames = func;
-            cd = new Condition(cond);        // full clone, we can now modify it.
-            actionfile = file;
-            actioncorecontroller = cp;
-            applicationfolder = appfolder;
-            InitSub(events);
-        }
-
-        public virtual string ID() { return ""; }
-        public virtual void InitSub(List<string> events) { }
-        public virtual void EventChanged() { }
-        public virtual void UpdateProgramList(string[] proglist) { }
-
-        public virtual new void Dispose()
-        {
-            base.Dispose();
-        }
-
-        public void RefreshIt()
-        {
-            RefreshEvent?.Invoke();
-        }
-
-        public void RemoveIt()
-        {
-            RemoveItem?.Invoke(this);
-        }
-    }
-
-    public class UserControlAction : UserControlAB
-    {
-        public ExtendedControls.ComboBoxCustom eventtype;
-        public ExtendedControls.ComboBoxCustom proglist;
-        public ExtendedControls.ButtonExt progedit;
-        public ExtendedControls.TextBoxBorder paras;
-
-        public override void InitSub(List<string> events)
-        {
-            eventtype = new ExtendedControls.ComboBoxCustom();
-            eventtype.Items.AddRange(events);
-            eventtype.Location = new Point(panelxmargin, panelymargin);
-            eventtype.Size = new Size(140, 24);
-            eventtype.DropDownHeight = 400;
-            eventtype.DropDownWidth = eventtype.Width * 3 / 2;
-            if ( cd.eventname !=null )
-                eventtype.SelectedItem = cd.eventname;
-            eventtype.SelectedIndexChanged += Actiontype_SelectedIndexChanged;
-
-            proglist = new ExtendedControls.ComboBoxCustom();
-            proglist.Items.Add("New");
-            proglist.Items.AddRange(actionfile.actionprogramlist.GetActionProgramList());
-            proglist.Location = new Point(eventtype.Right + 16, panelymargin);
-            proglist.Size = new Size(140, 24);
-            proglist.DropDownHeight = 400;
-            proglist.DropDownWidth = proglist.Width * 3 / 2;
-            if (cd.action != null)
-                proglist.Text = cd.action;
-            else
-                proglist.SelectedIndex = 0;
-
-            proglist.SelectedIndexChanged += Proglist_SelectedIndexChanged;
-
-            progedit = new ExtendedControls.ButtonExt();
-            progedit.Text = "P";
-            progedit.Location = new Point(proglist.Right + 8, panelymargin);
-            progedit.Size = new Size(24, 24);
-            progedit.Click += Progedit_Click;
-
-            paras = new ExtendedControls.TextBoxBorder();
-            paras.Text = (cd.actiondata != null) ? cd.actiondata : "";
-            paras.Location = new Point(progedit.Right + 8, panelymargin + 2);
-            paras.Size = new Size(96, 24);
-            paras.ReadOnly = true;
-            paras.Click += Paras_Click;
-
-            SuspendLayout();
-            Controls.Add(proglist);
-            Controls.Add(eventtype);
-            Controls.Add(progedit);
-            Controls.Add(paras);
-            ResumeLayout();
-        }
-
-        public override void UpdateProgramList(string[] progl)
-        {
-            proglist.Enabled = false;
-
-            string text = proglist.Text;
-
-            proglist.Items.Clear();
-            proglist.Items.Add("New");
-            proglist.Items.AddRange(progl);
-
-            if ( proglist.Items.Contains(text))
-                proglist.SelectedItem = text;
-
-            proglist.Enabled = true;
-        }
-
-        private void Paras_Click(object sender, EventArgs e)
-        {
-            ConditionVariables cond = new ConditionVariables();
-            string flag = "";
-
-            if (cd.actiondata != null)
-            {
-                cond.FromActionDataString(cd.actiondata, out flag);
-            }
-
-            ConditionVariablesForm avf = new ConditionVariablesForm();
-            avf.Init("Input parameters and flags to pass to program on run", this.Icon, cond, showone: true, showrefresh: true, showrefreshstate: flag.Equals(ConditionVariables.flagRunAtRefresh));
-
-            if (avf.ShowDialog(this) == DialogResult.OK)
-            {
-                cd.actiondata = paras.Text = avf.result.ToActionDataString(avf.result_refresh ? ConditionVariables.flagRunAtRefresh : "");
-            }
-        }
-
-        private void Progedit_Click(object sender, EventArgs e)
-        {
-            bool shift = ModifierKeys.HasFlag(Keys.Shift);
-
-            ActionProgram p = null;
-
-            if (proglist.SelectedIndex > 0)     // exclude NEW from checking for program
-                p = actionfile.actionprogramlist.Get(proglist.Text);
-
-            if (p != null && p.StoredInSubFile != null && shift)        // if we have a stored in sub file, but we shift hit, cancel it
-            {
-                if (ExtendedControls.MessageBoxTheme.Show(this, "Do you want to bring the file back into the main file", "WARNING", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
-                {
-                    p.CancelSubFileStorage();
-                    shift = false;
-                }
-                else
-                    return; // cancel, abort.
-            }
-
-            if (p != null && p.StoredInSubFile != null)
-            {
-                p.EditInEditor(p.StoredInSubFile);         // Edit in the editor.. this also updated the program steps held internally
-            }
-            else
-            {
-                string suggestedname = null;
-
-                if (p == null)        // if no program, create a new suggested name and clear any action data
-                {
-                    suggestedname = eventtype.Text;
-                    int n = 2;
-                    while (actionfile.actionprogramlist.GetActionProgramList().Contains(suggestedname))
-                    {
-                        suggestedname = eventtype.Text + "_" + n.ToString(System.Globalization.CultureInfo.InvariantCulture);
-                        n++;
-                    }
-
-                    paras.Text = "";
-                }
-
-                ActionProgramEditForm apf = new ActionProgramEditForm();
-                apf.EditProgram += EditProgram;
-
-                // we init with a variable list based on the field names of the group (normally the event field names got by SetFieldNames)
-                // pass in the program if found, and its action data.
-
-                apf.Init("Action program " , this.Icon, actioncorecontroller, applicationfolder, onAdditionalNames(eventtype.Text), actionfile.name , p, actionfile.actionprogramlist.GetActionProgramList(), suggestedname, ModifierKeys.HasFlag(Keys.Shift));
-
-                DialogResult res = apf.ShowDialog();
-
-                if (res == DialogResult.OK)
-                {
-                    ActionProgram np = apf.GetProgram();
-                    actionfile.actionprogramlist.AddOrChange(np);                // replaces or adds (if its a new name) same as rename
-                    cd.action = np.Name;
-                    RefreshIt();
-                    proglist.Enabled = false;
-                    proglist.SelectedItem = np.Name;
-                    proglist.Enabled = true;
-                }
-                else if (res == DialogResult.Abort)   // delete
-                {
-                    ActionProgram np2 = apf.GetProgram();
-                    actionfile.actionprogramlist.Delete(np2.Name);
-                    cd.action = "";
-                    RefreshIt();
-                }
-            }
-        }
-
-        private void EditProgram(string s)  // Callback by APF to ask to edit another program..
-        {
-            if (!actionfile.actionprogramlist.EditProgram(s, actionfile.name, actioncorecontroller, applicationfolder))
-                ExtendedControls.MessageBoxTheme.Show(this, "Unknown program or not in this file " + s);
-        }
-
-        private void Actiontype_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            cd.eventname = eventtype.Text;
-            EventChanged();
-        }
-
-        private void Proglist_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            if (proglist.Enabled && proglist.SelectedIndex == 0)   // if selected NEW.
-            {
-                Progedit_Click(null, null);
-            }
-            else
-                cd.action = proglist.Text;      // set program selected
-        }
-
-        public new void Dispose()
-        {
-            Controls.Clear();
-            proglist.Dispose();
-            eventtype.Dispose();
-            progedit.Dispose();
-            paras.Dispose();
-            base.Dispose();
-        }
-    }
-
-
-    public class UserControlProgram : UserControlAction
-    {
-        public ExtendedControls.TextBoxBorder condition;
-        public ExtendedControls.ButtonExt delete;
-
-        public override string ID() { return eventtype.Text.Length > 0 ? eventtype.Text : "Action not set"; }
-
-        public override void InitSub(List<string> events)
-        {
-            base.InitSub(events);
-
-            condition = new ExtendedControls.TextBoxBorder();
-            condition.Text = cd.ToString();
-            condition.Location = new Point(paras.Right + 16, panelymargin + 2);
-            condition.Size = new Size(200, 24);
-            condition.ReadOnly = true;
-            condition.Click += Condition_Click;
-            condition.Tag = this;
-
-            delete = new ExtendedControls.ButtonExt();
-            delete.Text = "X";
-            delete.Location = new Point(condition.Right + 8, panelymargin);
-            delete.Size = new Size(24, 24);
-            delete.Click += Delete_Click;
-
-            SuspendLayout();
-            Controls.Add(condition);
-            Controls.Add(delete);
-            ResumeLayout();
-        }
-
-        private void Condition_Click(object sender, EventArgs e)
-        {
-            ConditionFilterForm frm = new ConditionFilterForm();
-
-            frm.InitCondition("Action condition", this.Icon, onAdditionalNames(eventtype.Text), cd);
-            frm.TopMost = this.FindForm().TopMost;
-            if (frm.ShowDialog(this.FindForm()) == DialogResult.OK)
-            {
-                Condition res = frm.result.Get(0);
-                if (res != null)
-                {
-                    cd.fields = res.fields;
-                    cd.innercondition = res.innercondition;
-                }
-                else
-                    cd.fields = null;
-
-                condition.Text = cd.ToString();
-            }
-        }
-
-        private void Delete_Click(object sender, EventArgs e)
-        {
-            RemoveIt();
-        }
-
-        public new void Dispose()
-        {
-            condition.Dispose();
-            delete.Dispose();
-            base.Dispose();
-        }
-
-        public override void EventChanged()     // helpers to set initial conditions
-        {
-            if (cd.AlwaysTrue())
-            {
-                if (cd.eventname == "onKeyPress")
-                    cd.Set(new ConditionEntry("KeyPress", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onTimer")
-                    cd.Set(new ConditionEntry("TimerName", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onPopUp" || cd.eventname == "onPopDown")
-                    cd.Set(new ConditionEntry("PopOutName", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onTabChange")
-                    cd.Set(new ConditionEntry("TabName", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onPanelChange")
-                    cd.Set(new ConditionEntry("PopOutName", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onEliteInput" || cd.eventname == "onEliteInputOff" )
-                    cd.Set(new ConditionEntry("Binding", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onMenuItem")
-                    cd.Set(new ConditionEntry("MenuName", ConditionEntry.MatchType.Equals, "?"));
-                else if (cd.eventname == "onSayStarted" || cd.eventname == "onSayFinished" || cd.eventname == "onPlayStarted" || cd.eventname == "onPlayFinished" )
-                    cd.Set(new ConditionEntry("EventName", ConditionEntry.MatchType.Equals, "?"));
-
-                condition.Text = cd.ToString();
-            }
-        }
     }
 }
