+++ ActionLanguage/ActionLanguage.csproj
@@ -46,6 +46,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ActionEditing\ActionPackEditBase.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="ActionEditing\ActionPackEditEventProgramCondition.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
     <Compile Include="ActionsCoreCmds\ActionPlay.cs" />
     <Compile Include="ActionsCoreCmds\ActionPragma.cs" />
     <Compile Include="ActionsCoreCmds\ActionPrint.cs" />
