+++ ExtendedControls/DraggableForm.cs
@@ -1,0 +1,106 @@
+﻿using BaseUtils.Win32Constants;
+using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace ExtendedControls
+{
+    public class DraggableForm : Form
+    {
+        public void DragMoveMode(Control sender)  
+        {
+            sender.Capture = false;
+            SendMessage(WM.NCLBUTTONDOWN, (IntPtr)HT.CAPTION, IntPtr.Zero);
+        }
+
+        // Mono compatibility
+        private bool _window_dragging = false;
+        private Point _window_dragMousePos = Point.Empty;
+        private Point _window_dragWindowPos = Point.Empty;
+
+        private IntPtr SendMessage(int msg, IntPtr wparam, IntPtr lparam)
+        {
+            Message message = Message.Create(this.Handle, msg, wparam, lparam);
+            this.WndProc(ref message);
+            return message.Result;
+        }
+
+        protected override void WndProc(ref Message m)
+        {
+            bool windowsborder = this.FormBorderStyle == FormBorderStyle.Sizable;
+            // Compatibility movement for Mono
+            if (m.Msg == WM.LBUTTONDOWN && (int)m.WParam == 1 && !windowsborder)
+            {
+                int x = unchecked((short)((uint)m.LParam & 0xFFFF));
+                int y = unchecked((short)((uint)m.LParam >> 16));
+                _window_dragMousePos = new Point(x, y);
+                _window_dragWindowPos = this.Location;
+                _window_dragging = true;
+                m.Result = IntPtr.Zero;
+                this.Capture = true;
+            }
+            else if (m.Msg == WM.MOUSEMOVE && (int)m.WParam == 1 && _window_dragging)
+            {
+                int x = unchecked((short)((uint)m.LParam & 0xFFFF));
+                int y = unchecked((short)((uint)m.LParam >> 16));
+                Point delta = new Point(x - _window_dragMousePos.X, y - _window_dragMousePos.Y);
+                _window_dragWindowPos = new Point(_window_dragWindowPos.X + delta.X, _window_dragWindowPos.Y + delta.Y);
+                this.Location = _window_dragWindowPos;
+                this.Update();
+                m.Result = IntPtr.Zero;
+            }
+            else if (m.Msg == WM.LBUTTONUP)
+            {
+                _window_dragging = false;
+                _window_dragMousePos = Point.Empty;
+                _window_dragWindowPos = Point.Empty;
+                m.Result = IntPtr.Zero;
+                this.Capture = false;
+            }
+            // Windows honours NCHITTEST; Mono does not
+            else if (m.Msg == WM.NCHITTEST)
+            {
+                base.WndProc(ref m);
+
+                if ((int)m.Result == HT.CLIENT)
+                {
+                    int x = unchecked((short)((uint)m.LParam & 0xFFFF));
+                    int y = unchecked((short)((uint)m.LParam >> 16));
+                    Point p = PointToClient(new Point(x, y));
+
+                    StatusStripCustom statstrip = Controls.OfType<StatusStripCustom>().First();
+
+                    if (p.X > this.ClientSize.Width - statstrip.Height && p.Y > this.ClientSize.Height - statstrip.Height)
+                    {
+                        m.Result = (IntPtr)HT.BOTTOMRIGHT;
+                    }
+                    else if (p.Y > this.ClientSize.Height - statstrip.Height)
+                    {
+                        m.Result = (IntPtr)HT.BOTTOM;
+                    }
+                    else if (p.X > this.ClientSize.Width - 5)       // 5 is generous.. really only a few pixels gets thru before the subwindows grabs them
+                    {
+                        m.Result = (IntPtr)HT.RIGHT;
+                    }
+                    else if (p.X < 5)
+                    {
+                        m.Result = (IntPtr)HT.LEFT;
+                    }
+                    else if (!windowsborder)
+                    {
+                        m.Result = (IntPtr)HT.CAPTION;
+                    }
+                }
+            }
+            else
+            {
+                base.WndProc(ref m);
+            }
+        }
+
+    }
+}
