+++ ExtendedControls/ExtendedControls.csproj
@@ -61,6 +61,9 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="BitMapHelpers.cs" />
+    <Compile Include="DraggableForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="ThemeStandardEditor.cs">
       <SubType>Form</SubType>
     </Compile>
